+++ src/main/java/igentuman/nc/recipes/processors/DecayHastenerRecipe.java
@@ -48,7 +48,7 @@ public class DecayHastenerRecipe extends ItemStackToItemStackRecipe {
         return new ItemStack(NCProcessors.PROCESSORS.get(ID).get());
     }
 
-    public int getEnergy() {
+    public double getEnergy() {
         return Processors.all().get(ID).config().getPower();
     }
