+++ src/main/java/igentuman/nc/recipes/processors/PressurizerRecipe.java
@@ -46,7 +46,7 @@ public class PressurizerRecipe extends ItemStackToItemStackRecipe {
         return new ItemStack(NCProcessors.PROCESSORS.get(ID).get());
     }
 
-    public int getEnergy() {
+    public double getEnergy() {
         return Processors.all().get(ID).config().getPower();
     }
